[Fix] 공지사항 수정, 삭제에 이미지 추가
수정, 삭제 시에 해당 이미지도 같이 업데이트 되도록 코드 추가함

Related to: #4

+++ src/main/java/com/midnear/midnearshopping/domain/dto/notice/NoticeImagesDto.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @AllArgsConstructor
 @Builder
 public class NoticeImagesDto {
-    private Integer noticeImageId;
+    private Long noticeImageId;
 
     private String imageUrl;
 
@@ -21,7 +21,7 @@ public class NoticeImagesDto {
 
     private Date imageCreationDate;
 
-    private int noticeId;
+    private Long noticeId;
     public static NoticeImagesDto toDto(NoticeImagesVo noticeImagesVo) {
         return NoticeImagesDto.builder()
                 .noticeImageId(noticeImagesVo.getNoticeImageId())
